refactor: rearrange folder structure
NMN-148
