+++ app/(panels)/Control.tsx
